+++ csfunctions/actions/start_workflow.py
@@ -2,14 +2,8 @@ from typing import Literal
 
 from pydantic import BaseModel, Field
 
+from ..subject import Subject
 from .base import ActionNames, BaseAction
-
-
-class Subject(BaseModel):
-    subject_id: str = Field(..., description="ID of the subject, eg. a role name or personalnummer")
-    subject_type: Literal["Person", "PCS Role", "Common Role"] = Field(
-        ..., description="Type of the subject: Person, PCS Role or Common Role"
-    )
 
 
 class TaskConfiguration(BaseModel):